requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
Pillow==10.1.0
numpy<2
pyzbar==0.1.9
tkinter
